requirements.txt
fastapi
uvicorn
sqlalchemy
pytesseract
pillow
python-multipart
